+++ package.json
@@ -17,6 +17,7 @@
     "babel-core": "^6.26.0",
     "css-loader": "^0.28.11",
     "file-loader": "^1.1.11",
+    "font-awesome": "^4.7.0",
     "json-loader": "^0.5.7",
     "node-sass": "^4.9.0",
     "react": "^15.2.1",
