+++ src/app/components/NavigationBar/NavigationBar.scss
@@ -1,0 +1,13 @@
+#navBarContainer {
+    background-color: #24292E;
+    border: none;
+    border-radius: 0px;
+    height: fit-content;
+    width: 100%;
+    position: fixed;
+    vertical-align: middle;
+    a strong {
+        color: white;
+    }
+
+}
