+++ src/app/container/App.scss
@@ -1,0 +1,9 @@
+ #AppSection {
+    #headerSection{
+        height: 10%;
+    }
+    #bodySection {
+        height:90%;
+    }
+ }
+
