+++ src/app/index.css
@@ -4,7 +4,7 @@
     height: 100%;
     font-size: 16px;
     font-family: 'Calibri';
-    background: #f2f3f4;
+    background: #2B3137;
   }
   .no-margin {
     margin: 0px;
