+++ src/app/index.js
@@ -4,6 +4,8 @@ import { Provider } from 'react-redux';
 import { Router, Route, IndexRoute, browserHistory } from 'react-router';
 import store from './store';
 import App from './container/App.jsx';
+import './index.css';
+import 'font-awesome/css/font-awesome.css';
 
 render(
 <Provider store={store}>
